Refactor correlation tests and add new data tests - Remove unused imports and redundant tests from `test_correlation.py` - Create `test_data.py` with tests for `load_airline()` and `load_macrodata()` - Enhance error handling tests in `test_errors.py`
