+++ src/tests/test_correlation.py
@@ -10,12 +10,8 @@
 # ---------------------------------------------------------------------------- #
 
 
-# ## Python StdLib Imports ----
-from unittest.mock import patch
-
 # ## Python Third Party Imports ----
 import numpy as np
-import pandas as pd
 from pytest import raises
 from statsmodels import api as sm
 from statsmodels.stats.api import (
@@ -28,14 +24,11 @@ from statsmodels.tsa.stattools import (
     ccf as st_ccf,
     pacf as st_pacf,
 )
-from typeguard import TypeCheckError
 
 # ## Local First Party Imports ----
 from tests.setup import BaseTester
 from ts_stat_tests.algorithms.correlation import acf, bglm, ccf, lb, lm, pacf
 from ts_stat_tests.tests.correlation import correlation, is_correlated
-from ts_stat_tests.utils.data import get_uniform_data, load_airline, load_macrodata
-from ts_stat_tests.utils.errors import assert_almost_equal, is_almost_equal
 
 
 # ---------------------------------------------------------------------------- #
@@ -209,52 +202,3 @@ class TestCorrelation(BaseTester):
         res_noise = is_correlated(self.data_noise, algorithm="lb", lags=[10])
         # It's noise, so it should be False (stationary/non-correlated)
         assert isinstance(res_noise["result"], bool)
-
-    def test_load_airline_type_error(self) -> None:
-        mock_df = pd.DataFrame({"A": [1, 2], "B": [3, 4]})
-        load_airline.cache_clear()
-        with patch("pandas.read_csv", return_value=mock_df):
-            with raises(TypeError, match="Expected a pandas Series from the data source."):
-                load_airline()
-        load_airline.cache_clear()
-
-    def test_load_macrodata_type_error(self) -> None:
-        load_macrodata.cache_clear()
-        with patch("pandas.read_csv", return_value=pd.Series(dtype=float)):
-            with raises(TypeCheckError) as e:
-                load_macrodata()
-                assert "value assigned to data (pandas.core.series.Series)" in str(e.value)
-                assert "is not an instance of" in str(e.value)
-                assert "pandas.core.frame.DataFrame" in str(e.value)
-
-        load_macrodata.cache_clear()
-
-    def test_get_uniform_data(self) -> None:
-        res = get_uniform_data(42)
-        assert isinstance(res, np.ndarray)
-        assert res.shape == (1000,)
-
-    def test_is_almost_equal_value_error(self) -> None:
-        with raises(ValueError, match="Specify `delta` or `places`, not both."):
-            is_almost_equal(1.0, 1.1, places=7, delta=0.1)
-
-    def test_is_almost_equal_with_delta(self) -> None:
-        assert is_almost_equal(1.0, 1.05, delta=0.1) is True
-        assert is_almost_equal(1.0, 1.2, delta=0.1) is False
-
-    def test_assert_almost_equal_failures_1(self) -> None:
-        with raises(AssertionError) as e:
-            assert_almost_equal(1.0, 1.2, delta=0.1)
-            assert "1.0 != 1.2" in str(e.value)
-            assert "delta=0.1" in str(e.value)
-
-    def test_assert_almost_equal_failures_2(self) -> None:
-        with raises(AssertionError) as e:
-            assert_almost_equal(1.0, 1.0000001, places=10)
-            assert "1.0 != 1.0000001" in str(e.value)
-            assert "places=10" in str(e.value)
-
-    def test_assert_almost_equal_failures_3(self) -> None:
-        with raises(AssertionError) as e:
-            assert_almost_equal(1.0, 2.0, msg="Custom error message")
-            assert "Custom error message" in str(e.value)
