+++ src/tests/test_errors.py
@@ -1,3 +1,15 @@
+# ---------------------------------------------------------------------------- #
+#                                                                              #
+#    Setup                                                                  ####
+#                                                                              #
+# ---------------------------------------------------------------------------- #
+
+
+# ---------------------------------------------------------------------------- #
+# Imports                                                                   ####
+# ---------------------------------------------------------------------------- #
+
+
 # ## Python StdLib Imports ----
 import unittest
 
@@ -10,6 +22,18 @@ from ts_stat_tests.utils.errors import (
     generate_error_message,
     is_almost_equal,
 )
+
+
+# ---------------------------------------------------------------------------- #
+#                                                                              #
+#    Testing                                                                ####
+#                                                                              #
+# ---------------------------------------------------------------------------- #
+
+
+# ---------------------------------------------------------------------------- #
+# Tests                                                                     ####
+# ---------------------------------------------------------------------------- #
 
 
 class TestErrors(unittest.TestCase):
@@ -29,10 +53,7 @@ class TestErrors(unittest.TestCase):
         assert is_almost_equal(1.0, 1.05, delta=0.1)
         assert not is_almost_equal(1.0, 1.05, delta=0.01)
 
-        with raises(ValueError):
-            is_almost_equal(1.0, 1.1, places=7, delta=0.1)
-
-    def test_assert_almost_equal(self):
+    def test_assert_almost_equal(self) -> None:
         assert_almost_equal(1.0, 1.0)
         assert_almost_equal(1.0, 1.00000001, places=7)
         assert_almost_equal(1.0, 1.05, delta=0.1)
@@ -45,3 +66,28 @@ class TestErrors(unittest.TestCase):
 
         with raises(AssertionError, match="custom"):
             assert_almost_equal(1.0, 1.1, msg="custom")
+
+    def test_is_almost_equal_value_error(self) -> None:
+        with raises(ValueError, match="Specify `delta` or `places`, not both."):
+            is_almost_equal(1.0, 1.1, places=7, delta=0.1)
+
+    def test_is_almost_equal_with_delta(self) -> None:
+        assert is_almost_equal(1.0, 1.05, delta=0.1) is True
+        assert is_almost_equal(1.0, 1.2, delta=0.1) is False
+
+    def test_assert_almost_equal_failures_1(self) -> None:
+        with raises(AssertionError) as e:
+            assert_almost_equal(1.0, 1.2, delta=0.1)
+        assert "1.0 != 1.2" in str(e.value)
+        assert "delta=0.1" in str(e.value)
+
+    def test_assert_almost_equal_failures_2(self) -> None:
+        with raises(AssertionError) as e:
+            assert_almost_equal(1.0, 1.0000001, places=10)
+        assert "1.0 != 1.0000001" in str(e.value)
+        assert "places=10" in str(e.value)
+
+    def test_assert_almost_equal_failures_3(self) -> None:
+        with raises(AssertionError) as e:
+            assert_almost_equal(1.0, 2.0, msg="Custom error message")
+        assert "Custom error message" in str(e.value)
